Improved `ObjectBoxToLibkiwixMigratorTest` for clearing the previous saved bookmarks.
* Refresh the data before checking the loaded data in DownloadTest.
* Increasing the Retry count for loading data and downloading start to properly work with low internet connection.

+++ app/src/androidTest/java/org/kiwix/kiwixmobile/ObjectBoxToLibkiwixMigratorTest.kt
@@ -119,6 +119,8 @@ class ObjectBoxToLibkiwixMigratorTest : BaseActivityTest() {
       )
     }
     box = boxStore.boxFor(BookmarkEntity::class.java)
+    // clear the data before running the test case
+    clearBookmarks()
 
     // add a file in fileSystem because we need to actual file path for making object of Archive.
     val loadFileStream =
@@ -139,9 +141,6 @@ class ObjectBoxToLibkiwixMigratorTest : BaseActivityTest() {
         }
       }
     }
-
-    // clear the data before running the test case
-    clearBookmarks()
   }
 
   @Test
@@ -310,7 +309,9 @@ class ObjectBoxToLibkiwixMigratorTest : BaseActivityTest() {
         .blockingFirst() as List<LibkiwixBookmarkItem>
     )
     box.removeAll()
-    zimFile.delete() // delete the temp ZIM file to free up the memory
+    if (::zimFile.isInitialized) {
+      zimFile.delete() // delete the temp ZIM file to free up the memory
+    }
   }
 
   @After
